[CHORE] Backup Config class 분리
- 공통기능 빈 등록 Config와 backup Config 분리
